Update load-balancer-limits.md
upon attempting to update cloud formation template that updates a rule with more then 6 wildcards in it, it rolled back with an error stating the max wildcards you can use is 6. I updated the template again using the proposed limit and it is working. The documentation should change to reflect this new limit