Update http2 sample request log for ALB
It was incorrect. The following fields are actually enclosed in double quotes, but in the example are not enclosed in double quotes:

target:port_list
target_status_code_list
classification
classification_reason
Issue #, if available:

Description of changes:

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.